Rewrite to be more fully side-by-side
